Fix all mypy issues, Fix all pylint issue, Refactor code
