+++ ariadne_graphql_modules/next/graphql_input/__init__.py
@@ -1,0 +1,8 @@
+from .input_type import GraphQLInput
+from .input_model import GraphQLInputModel
+
+
+__all__ = [
+    "GraphQLInput",
+    "GraphQLInputModel",
+]
