+++ ariadne_graphql_modules/next/graphql_interface/__init__.py
@@ -1,0 +1,8 @@
+from .interface_type import GraphQLInterface
+from .interface_model import GraphQLInterfaceModel
+
+
+__all__ = [
+    "GraphQLInterface",
+    "GraphQLInterfaceModel",
+]
