+++ ariadne_graphql_modules/next/graphql_interface/interface_type.py
@@ -1,101 +1,56 @@
-from dataclasses import dataclass
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    Optional,
-    cast,
-)
+from typing import Any, Dict, List, cast
 
-from ariadne import SubscriptionType
-from ariadne.types import Resolver, Subscriber
+from ariadne.types import Resolver
 from graphql import (
     FieldDefinitionNode,
-    GraphQLField,
-    GraphQLObjectType,
-    GraphQLSchema,
     InputValueDefinitionNode,
+    InterfaceTypeDefinitionNode,
     NameNode,
     NamedTypeNode,
-    ObjectTypeDefinitionNode,
+    StringValueNode,
 )
 
-
-from .value import get_value_node
-
-from .objecttype import (
-    GraphQLObject,
-    GraphQLObjectResolver,
-    GraphQLObjectSource,
+from ...utils import parse_definition
+from ..base import GraphQLMetadata
+from ..description import get_description_node
+from ..graphql_object import GraphQLObject, GraphQLObjectResolver
+from ..graphql_object.object_type import get_graphql_object_data
+from ..graphql_object.utils import (
     get_field_args_from_resolver,
-    get_field_args_from_subscriber,
     get_field_args_out_names,
     get_field_node_from_obj_field,
-    get_graphql_object_data,
-    object_subscriber,
     update_field_args_options,
 )
-
-from ..utils import parse_definition
-from .base import GraphQLMetadata, GraphQLModel
-from .description import get_description_node
+from ..value import get_value_node
+from .interface_model import GraphQLInterfaceModel
 
 
-class GraphQLSubscription(GraphQLObject):
+class GraphQLInterface(GraphQLObject):
     __abstract__: bool = True
-    __valid_type__ = ObjectTypeDefinitionNode
+    __valid_type__ = InterfaceTypeDefinitionNode
 
     @classmethod
-    def __get_graphql_model_with_schema__(cls, *_) -> "GraphQLSubscriptionModel":
+    def __get_graphql_model_with_schema__(cls) -> "GraphQLInterfaceModel":
         definition = cast(
-            ObjectTypeDefinitionNode,
-            parse_definition(ObjectTypeDefinitionNode, cls.__schema__),
+            InterfaceTypeDefinitionNode,
+            parse_definition(InterfaceTypeDefinitionNode, cls.__schema__),
         )
 
-        descriptions: Dict[str, str] = {}
-        args_descriptions: Dict[str, Dict[str, str]] = {}
+        descriptions: Dict[str, StringValueNode] = {}
+        args_descriptions: Dict[str, Dict[str, StringValueNode]] = {}
         args_defaults: Dict[str, Dict[str, Any]] = {}
         resolvers: Dict[str, Resolver] = {}
-        subscribers: Dict[str, Subscriber] = {}
         out_names: Dict[str, Dict[str, str]] = {}
 
         for attr_name in dir(cls):
             cls_attr = getattr(cls, attr_name)
             if isinstance(cls_attr, GraphQLObjectResolver):
                 resolvers[cls_attr.field] = cls_attr.resolver
-                if cls_attr.description:
-                    descriptions[cls_attr.field] = get_description_node(
-                        cls_attr.description
-                    )
+                description_node = get_description_node(cls_attr.description)
+                if description_node:
+                    descriptions[cls_attr.field] = description_node
 
                 field_args = get_field_args_from_resolver(cls_attr.resolver)
-                if field_args:
-                    args_descriptions[cls_attr.field] = {}
-                    args_defaults[cls_attr.field] = {}
-
-                    final_args = update_field_args_options(field_args, cls_attr.args)
-
-                    for arg_name, arg_options in final_args.items():
-                        arg_description = get_description_node(arg_options.description)
-                        if arg_description:
-                            args_descriptions[cls_attr.field][
-                                arg_name
-                            ] = arg_description
-
-                        arg_default = arg_options.default_value
-                        if arg_default is not None:
-                            args_defaults[cls_attr.field][arg_name] = get_value_node(
-                                arg_default
-                            )
-            if isinstance(cls_attr, GraphQLObjectSource):
-                subscribers[cls_attr.field] = cls_attr.subscriber
-                if cls_attr.description:
-                    descriptions[cls_attr.field] = get_description_node(
-                        cls_attr.description
-                    )
-
-                field_args = get_field_args_from_subscriber(cls_attr.subscriber)
                 if field_args:
                     args_descriptions[cls_attr.field] = {}
                     args_defaults[cls_attr.field] = {}
@@ -149,17 +104,16 @@ class GraphQLSubscription(GraphQLObject):
                 )
             )
 
-        return GraphQLSubscriptionModel(
+        return GraphQLInterfaceModel(
             name=definition.name.value,
-            ast_type=ObjectTypeDefinitionNode,
-            ast=ObjectTypeDefinitionNode(
+            ast_type=InterfaceTypeDefinitionNode,
+            ast=InterfaceTypeDefinitionNode(
                 name=NameNode(value=definition.name.value),
                 fields=tuple(fields),
                 interfaces=definition.interfaces,
             ),
             resolve_type=cls.resolve_type,
             resolvers=resolvers,
-            subscribers=subscribers,
             aliases=getattr(cls, "__aliases__", {}),
             out_names=out_names,
         )
@@ -167,40 +121,37 @@ class GraphQLSubscription(GraphQLObject):
     @classmethod
     def __get_graphql_model_without_schema__(
         cls, metadata: GraphQLMetadata, name: str
-    ) -> "GraphQLSubscriptionModel":
+    ) -> "GraphQLInterfaceModel":
         type_data = get_graphql_object_data(metadata, cls)
         type_aliases = getattr(cls, "__aliases__", None) or {}
 
         fields_ast: List[FieldDefinitionNode] = []
         resolvers: Dict[str, Resolver] = {}
-        subscribers: Dict[str, Subscriber] = {}
         aliases: Dict[str, str] = {}
         out_names: Dict[str, Dict[str, str]] = {}
 
         for attr_name, field in type_data.fields.items():
             fields_ast.append(get_field_node_from_obj_field(cls, metadata, field))
-            if attr_name in type_aliases:
+
+            if attr_name in type_aliases and field.name:
                 aliases[field.name] = type_aliases[attr_name]
-            elif attr_name != field.name and not field.resolver:
+            elif field.name and attr_name != field.name and not field.resolver:
                 aliases[field.name] = attr_name
 
-            if field.resolver:
+            if field.resolver and field.name:
                 resolvers[field.name] = field.resolver
 
-            if field.subscriber:
-                subscribers[field.name] = field.subscriber
-
-            if field.args:
+            if field.args and field.name:
                 out_names[field.name] = get_field_args_out_names(field.args)
 
         interfaces_ast: List[NamedTypeNode] = []
         for interface_name in type_data.interfaces:
             interfaces_ast.append(NamedTypeNode(name=NameNode(value=interface_name)))
 
-        return GraphQLSubscriptionModel(
+        return GraphQLInterfaceModel(
             name=name,
-            ast_type=ObjectTypeDefinitionNode,
-            ast=ObjectTypeDefinitionNode(
+            ast_type=InterfaceTypeDefinitionNode,
+            ast=InterfaceTypeDefinitionNode(
                 name=NameNode(value=name),
                 description=get_description_node(
                     getattr(cls, "__description__", None),
@@ -212,54 +163,13 @@ class GraphQLSubscription(GraphQLObject):
             resolvers=resolvers,
             aliases=aliases,
             out_names=out_names,
-            subscribers=subscribers,
         )
 
     @staticmethod
     def resolve_type(obj: Any, *_) -> str:
-        if isinstance(obj, GraphQLSubscription):
+        if isinstance(obj, GraphQLInterface):
             return obj.__get_graphql_name__()
 
         raise ValueError(
             f"Cannot resolve GraphQL type {obj} for object of type '{type(obj).__name__}'."
         )
-
-    @staticmethod
-    def source(
-        field: str,
-        graphql_type: Optional[Any] = None,
-        args: Optional[Dict[str, dict]] = None,
-        description: Optional[str] = None,
-    ):
-        """Shortcut for object_resolver()"""
-        return object_subscriber(
-            args=args,
-            field=field,
-            graphql_type=graphql_type,
-            description=description,
-        )
-
-
-@dataclass(frozen=True)
-class GraphQLSubscriptionModel(GraphQLModel):
-    resolvers: Dict[str, Resolver]
-    resolve_type: Callable[[Any], Any]
-    out_names: Dict[str, Dict[str, str]]
-    aliases: Dict[str, str]
-    subscribers: Dict[str, Subscriber]
-
-    def bind_to_schema(self, schema: GraphQLSchema):
-        bindable = SubscriptionType()
-        for field, resolver in self.resolvers.items():
-            bindable.set_field(field, resolver)
-        for alias, target in self.aliases.items():
-            bindable.set_alias(alias, target)
-        for source, generator in self.subscribers.items():
-            bindable.set_source(source, generator)
-        bindable.bind_to_schema(schema)
-
-        graphql_type = cast(GraphQLObjectType, schema.get_type(self.name))
-        for field_name, field_out_names in self.out_names.items():
-            graphql_field = cast(GraphQLField, graphql_type.fields[field_name])
-            for arg_name, out_name in field_out_names.items():
-                graphql_field.args[arg_name].out_name = out_name
