+++ ariadne_graphql_modules/next/graphql_scalar/__init__.py
@@ -1,0 +1,8 @@
+from .scalar_type import GraphQLScalar
+from .scalar_model import GraphQLScalarModel
+
+
+__all__ = [
+    "GraphQLScalar",
+    "GraphQLScalarModel",
+]
