+++ ariadne_graphql_modules/next/graphql_scalar/scalar_type.py
@@ -1,19 +1,21 @@
-from dataclasses import dataclass
-from typing import Any, Callable, Dict, Generic, Optional, Type, TypeVar, cast
+from typing import Any, Dict, Generic, Optional, TypeVar, cast
 
-from ariadne import ScalarType as ScalarTypeBindable
 from graphql import (
-    GraphQLSchema,
     NameNode,
     ScalarTypeDefinitionNode,
     ValueNode,
     value_from_ast_untyped,
 )
 
-from ..utils import parse_definition
-from .base import GraphQLMetadata, GraphQLModel, GraphQLType
-from .description import get_description_node
-from .validators import validate_description, validate_name
+from ..description import get_description_node
+from .scalar_model import GraphQLScalarModel
+
+from ...utils import parse_definition
+
+from .validators import validate_scalar_type_with_schema
+
+from ..base import GraphQLMetadata, GraphQLModel, GraphQLType
+
 
 T = TypeVar("T")
 
@@ -98,35 +100,3 @@ class GraphQLScalar(GraphQLType, Generic[T]):
 
     def unwrap(self) -> T:
         return self.wrapped_value
-
-
-@dataclass(frozen=True)
-class GraphQLScalarModel(GraphQLModel):
-    serialize: Callable[[Any], Any]
-    parse_value: Callable[[Any], Any]
-    parse_literal: Callable[[ValueNode, Optional[Dict[str, Any]]], Any]
-
-    def bind_to_schema(self, schema: GraphQLSchema):
-        bindable = ScalarTypeBindable(
-            self.name,
-            serializer=self.serialize,
-            value_parser=self.parse_value,
-            literal_parser=self.parse_literal,
-        )
-
-        bindable.bind_to_schema(schema)
-
-
-def validate_scalar_type_with_schema(cls: Type[GraphQLScalar]):
-    definition = parse_definition(cls.__name__, cls.__schema__)
-
-    if not isinstance(definition, ScalarTypeDefinitionNode):
-        raise ValueError(
-            f"Class '{cls.__name__}' defines '__schema__' attribute "
-            "with declaration for an invalid GraphQL type. "
-            f"('{definition.__class__.__name__}' != "
-            f"'{ScalarTypeDefinitionNode.__name__}')"
-        )
-
-    validate_name(cls, definition)
-    validate_description(cls, definition)
