+++ ariadne_graphql_modules/next/graphql_subscription/subscription_type.py
@@ -1,69 +1,89 @@
-from dataclasses import dataclass
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    cast,
-)
+from typing import Any, Dict, List, Optional, cast
 
-from ariadne import InterfaceType
-from ariadne.types import Resolver
 from graphql import (
     FieldDefinitionNode,
-    GraphQLField,
-    GraphQLObjectType,
-    GraphQLSchema,
     InputValueDefinitionNode,
     NameNode,
     NamedTypeNode,
-    InterfaceTypeDefinitionNode,
+    ObjectTypeDefinitionNode,
+    StringValueNode,
 )
 
-from .value import get_value_node
+from ariadne.types import Resolver, Subscriber
 
-from .objecttype import (
+
+from ...utils import parse_definition
+from ..base import GraphQLMetadata, GraphQLModel
+from ..description import get_description_node
+from ..graphql_object import (
     GraphQLObject,
     GraphQLObjectResolver,
+    GraphQLObjectSource,
+    GraphQLObjectFieldArg,
     get_field_args_from_resolver,
+    get_field_args_from_subscriber,
     get_field_args_out_names,
     get_field_node_from_obj_field,
     get_graphql_object_data,
+    object_subscriber,
     update_field_args_options,
 )
-
-from ..utils import parse_definition
-from .base import GraphQLMetadata, GraphQLModel
-from .description import get_description_node
+from ..value import get_value_node
+from .subscription_model import GraphQLSubscriptionModel
 
 
-class GraphQLInterface(GraphQLObject):
+class GraphQLSubscription(GraphQLObject):
     __abstract__: bool = True
-    __valid_type__ = InterfaceTypeDefinitionNode
+    __valid_type__ = ObjectTypeDefinitionNode
 
     @classmethod
-    def __get_graphql_model_with_schema__(cls) -> "GraphQLInterfaceModel":
+    def __get_graphql_model_with_schema__(cls) -> "GraphQLModel":
         definition = cast(
-            InterfaceTypeDefinitionNode,
-            parse_definition(InterfaceTypeDefinitionNode, cls.__schema__),
+            ObjectTypeDefinitionNode,
+            parse_definition(ObjectTypeDefinitionNode, cls.__schema__),
         )
 
-        descriptions: Dict[str, str] = {}
-        args_descriptions: Dict[str, Dict[str, str]] = {}
+        descriptions: Dict[str, StringValueNode] = {}
+        args_descriptions: Dict[str, Dict[str, StringValueNode]] = {}
         args_defaults: Dict[str, Dict[str, Any]] = {}
         resolvers: Dict[str, Resolver] = {}
+        subscribers: Dict[str, Subscriber] = {}
         out_names: Dict[str, Dict[str, str]] = {}
 
         for attr_name in dir(cls):
             cls_attr = getattr(cls, attr_name)
             if isinstance(cls_attr, GraphQLObjectResolver):
                 resolvers[cls_attr.field] = cls_attr.resolver
-                if cls_attr.description:
-                    descriptions[cls_attr.field] = get_description_node(
-                        cls_attr.description
-                    )
+                description_node = get_description_node(cls_attr.description)
+                if description_node:
+                    descriptions[cls_attr.field] = description_node
 
                 field_args = get_field_args_from_resolver(cls_attr.resolver)
+                if field_args:
+                    args_descriptions[cls_attr.field] = {}
+                    args_defaults[cls_attr.field] = {}
+
+                    final_args = update_field_args_options(field_args, cls_attr.args)
+
+                    for arg_name, arg_options in final_args.items():
+                        arg_description = get_description_node(arg_options.description)
+                        if arg_description:
+                            args_descriptions[cls_attr.field][
+                                arg_name
+                            ] = arg_description
+
+                        arg_default = arg_options.default_value
+                        if arg_default is not None:
+                            args_defaults[cls_attr.field][arg_name] = get_value_node(
+                                arg_default
+                            )
+            if isinstance(cls_attr, GraphQLObjectSource):
+                subscribers[cls_attr.field] = cls_attr.subscriber
+                description_node = get_description_node(cls_attr.description)
+                if description_node:
+                    descriptions[cls_attr.field] = description_node
+
+                field_args = get_field_args_from_subscriber(cls_attr.subscriber)
                 if field_args:
                     args_descriptions[cls_attr.field] = {}
                     args_defaults[cls_attr.field] = {}
@@ -117,16 +137,17 @@ class GraphQLInterface(GraphQLObject):
                 )
             )
 
-        return GraphQLInterfaceModel(
+        return GraphQLSubscriptionModel(
             name=definition.name.value,
-            ast_type=InterfaceTypeDefinitionNode,
-            ast=InterfaceTypeDefinitionNode(
+            ast_type=ObjectTypeDefinitionNode,
+            ast=ObjectTypeDefinitionNode(
                 name=NameNode(value=definition.name.value),
                 fields=tuple(fields),
                 interfaces=definition.interfaces,
             ),
             resolve_type=cls.resolve_type,
             resolvers=resolvers,
+            subscribers=subscribers,
             aliases=getattr(cls, "__aliases__", {}),
             out_names=out_names,
         )
@@ -134,25 +155,28 @@ class GraphQLInterface(GraphQLObject):
     @classmethod
     def __get_graphql_model_without_schema__(
         cls, metadata: GraphQLMetadata, name: str
-    ) -> "GraphQLInterfaceModel":
+    ) -> "GraphQLModel":
         type_data = get_graphql_object_data(metadata, cls)
         type_aliases = getattr(cls, "__aliases__", None) or {}
 
         fields_ast: List[FieldDefinitionNode] = []
         resolvers: Dict[str, Resolver] = {}
+        subscribers: Dict[str, Subscriber] = {}
         aliases: Dict[str, str] = {}
         out_names: Dict[str, Dict[str, str]] = {}
 
         for attr_name, field in type_data.fields.items():
             fields_ast.append(get_field_node_from_obj_field(cls, metadata, field))
-
-            if attr_name in type_aliases:
+            if attr_name in type_aliases and field.name:
                 aliases[field.name] = type_aliases[attr_name]
-            elif attr_name != field.name and not field.resolver:
+            elif field.name and attr_name != field.name and not field.resolver:
                 aliases[field.name] = attr_name
 
             if field.resolver and field.name:
                 resolvers[field.name] = field.resolver
+
+            if field.subscriber and field.name:
+                subscribers[field.name] = field.subscriber
 
             if field.args and field.name:
                 out_names[field.name] = get_field_args_out_names(field.args)
@@ -161,10 +185,10 @@ class GraphQLInterface(GraphQLObject):
         for interface_name in type_data.interfaces:
             interfaces_ast.append(NamedTypeNode(name=NameNode(value=interface_name)))
 
-        return GraphQLInterfaceModel(
+        return GraphQLSubscriptionModel(
             name=name,
-            ast_type=InterfaceTypeDefinitionNode,
-            ast=InterfaceTypeDefinitionNode(
+            ast_type=ObjectTypeDefinitionNode,
+            ast=ObjectTypeDefinitionNode(
                 name=NameNode(value=name),
                 description=get_description_node(
                     getattr(cls, "__description__", None),
@@ -176,36 +200,29 @@ class GraphQLInterface(GraphQLObject):
             resolvers=resolvers,
             aliases=aliases,
             out_names=out_names,
+            subscribers=subscribers,
         )
 
     @staticmethod
     def resolve_type(obj: Any, *_) -> str:
-        if isinstance(obj, GraphQLInterface):
+        if isinstance(obj, GraphQLSubscription):
             return obj.__get_graphql_name__()
 
         raise ValueError(
             f"Cannot resolve GraphQL type {obj} for object of type '{type(obj).__name__}'."
         )
 
-
-@dataclass(frozen=True)
-class GraphQLInterfaceModel(GraphQLModel):
-    resolvers: Dict[str, Resolver]
-    resolve_type: Callable[[Any], Any]
-    out_names: Dict[str, Dict[str, str]]
-    aliases: Dict[str, str]
-
-    def bind_to_schema(self, schema: GraphQLSchema):
-        bindable = InterfaceType(self.name, self.resolve_type)
-        for field, resolver in self.resolvers.items():
-            bindable.set_field(field, resolver)
-        for alias, target in self.aliases.items():
-            bindable.set_alias(alias, target)
-
-        bindable.bind_to_schema(schema)
-
-        graphql_type = cast(GraphQLObjectType, schema.get_type(self.name))
-        for field_name, field_out_names in self.out_names.items():
-            graphql_field = cast(GraphQLField, graphql_type.fields[field_name])
-            for arg_name, out_name in field_out_names.items():
-                graphql_field.args[arg_name].out_name = out_name
+    @staticmethod
+    def source(
+        field: str,
+        graphql_type: Optional[Any] = None,
+        args: Optional[Dict[str, GraphQLObjectFieldArg]] = None,
+        description: Optional[str] = None,
+    ):
+        """Shortcut for object_resolver()"""
+        return object_subscriber(
+            args=args,
+            field=field,
+            graphql_type=graphql_type,
+            description=description,
+        )
