+++ ariadne_graphql_modules/next/graphql_union/__init__.py
@@ -1,0 +1,8 @@
+from .union_type import GraphQLUnion
+from .union_model import GraphQLUnionModel
+
+
+__all__ = [
+    "GraphQLUnion",
+    "GraphQLUnionModel",
+]
